cy konfig ny port

+++ prisma/reset.ts
@@ -1,6 +1,8 @@
 import { db } from "./db";
 
 async function main() {
+  if (process.env.NODE_ENV !== "test") return;
+
   await db.fish.deleteMany({});
   //ta bort från alla tabeller
 }
